refactor: remove duplicate code
