+++ tests/unit/bff/v1/conftest.py
@@ -14,14 +14,6 @@ def base64_image_string(resources_folder_path: str) -> str:
 
 
 @pytest.fixture
-def base64_image_string(resources_folder_path: str) -> str:
-    with open(os.path.join(resources_folder_path, 'image', 'a.jpg'), 'rb') as f:
-        binary = f.read()
-        img_string = base64.b64encode(binary).decode('utf-8')
-    return img_string
-
-
-@pytest.fixture
 def sample_search_response_image() -> DocumentArray:
     result = DocumentArray([Document()])
     matches = DocumentArray([Document(uri='match')])
